Extract printTernaryOperator
